fix(db): simplify skill binary metadata migration
